--- zeyad/jsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    // ...
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-    // ...
-  }
-}
