--- zeyad/src/pages/Balance.jsx
@@ -1,96 +1,0 @@
-import axios from "axios";
-import { useEffect } from "react";
-import { useLocation } from "react-router-dom";
-
-function Balance() {
-  let location = useLocation().search;
-  location.replace("%20", " ");
-  console.log(location);
-
-  const fetchData = async () => {
-    const response = await axios.get(
-      `http://127.0.0.1:8000/api/v1/balance${location}&company=sample company`
-    );
-    console.log(response);
-  };
-
-  useEffect(() => {
-    fetchData();
-  }, []);
-
-  useEffect(() => {
-    fetchData();
-  }, []);
-
-  return (
-    <div className="overflow-x-auto rounded-lg border border-gray-200">
-      <table className="min-w-full divide-y-2 divide-gray-200 bg-white text-sm">
-        <thead className="">
-          <tr>
-            <th className="whitespace-nowrap py-2 font-medium text-gray-900">
-              Name
-            </th>
-            <th className="whitespace-nowrap py-2 font-medium text-gray-900">
-              Date of Birth
-            </th>
-            <th className="whitespace-nowrap px-4 py-2 font-medium text-gray-900">
-              Role
-            </th>
-            <th className="whitespace-nowrap px-4 py-2 font-medium text-gray-900">
-              Salary
-            </th>
-          </tr>
-        </thead>
-
-        <tbody className="divide-y divide-gray-200 text-center">
-          <tr>
-            <td className="whitespace-nowrap px-4 ml-4 py-2 font-medium text-gray-900">
-              John Doe
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              24/05/1995
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              Web Developer
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              $120,000
-            </td>
-          </tr>
-
-          <tr>
-            <td className="whitespace-nowrap px-4 py-2 font-medium text-gray-900">
-              Jane Doe
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              04/11/1980
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              Web Designer
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              $100,000
-            </td>
-          </tr>
-
-          <tr>
-            <td className="whitespace-nowrap px-4 py-2 font-medium text-gray-900">
-              Gary Barlow
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              24/05/1995
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              Singer
-            </td>
-            <td className="whitespace-nowrap px-4 py-2 text-gray-700">
-              $20,000
-            </td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-export default Balance;
