--- zeyad/src/pages/Compare.jsx
@@ -1,5 +1,0 @@
-function Compare() {
-  return <div>Compare</div>;
-}
-
-export default Compare;
